+++ Tests/NukePerformanceTests/ImageProcessingPerformanceTests.swift
@@ -2,11 +2,15 @@
 //
 // Copyright (c) 2015-2026 Alexander Grebenyuk (github.com/kean).
 
-import XCTest
+import Testing
+import Foundation
 import Nuke
 
-class ImageProcessingPerformanceTests: XCTestCase {
-    func testCreatingProcessorIdentifiers() {
+@Suite(.serialized)
+@MainActor
+struct ImageProcessingPerformanceTests {
+    @Test
+    func creatingProcessorIdentifiers() {
         let decompressor = ImageProcessors.Resize(size: CGSize(width: 1, height: 1), contentMode: .aspectFill, upscale: false)
 
         measure {
@@ -16,7 +20,8 @@ class ImageProcessingPerformanceTests: XCTestCase {
         }
     }
 
-    func testComparingTwoProcessorCompositions() {
+    @Test
+    func comparingTwoProcessorCompositions() {
         let lhs = ImageProcessors.Composition([MockImageProcessor(id: "123"), ImageProcessors.Resize(size: CGSize(width: 1, height: 1), contentMode: .aspectFill, upscale: false)])
         let rhs = ImageProcessors.Composition([MockImageProcessor(id: "124"), ImageProcessors.Resize(size: CGSize(width: 1, height: 1), contentMode: .aspectFill, upscale: false)])
 
@@ -29,7 +34,8 @@ class ImageProcessingPerformanceTests: XCTestCase {
         }
     }
 
-    func testImageDecoding() {
+    @Test
+    func imageDecoding() {
         let decoder = ImageDecoders.Default()
 
         let data = Test.data
@@ -42,8 +48,9 @@ class ImageProcessingPerformanceTests: XCTestCase {
 
     // MARK: Creating Thumbnails
 
-    func testResizeImage() throws {
-        let image = try XCTUnwrap(makeHighResolutionImage())
+    @Test
+    func resizeImage() throws {
+        let image = try #require(makeHighResolutionImage())
         let processor = ImageProcessors.Resize(size: CGSize(width: 64, height: 64), unit: .pixels)
 
         measure {
@@ -53,9 +60,10 @@ class ImageProcessingPerformanceTests: XCTestCase {
         }
     }
 
-    func testCreateThumbnail() throws {
-        let image = try XCTUnwrap(makeHighResolutionImage())
-        let data = try XCTUnwrap(ImageEncoders.ImageIO(type: .jpeg).encode(image))
+    @Test
+    func createThumbnail() throws {
+        let image = try #require(makeHighResolutionImage())
+        let data = try #require(ImageEncoders.ImageIO(type: .jpeg).encode(image))
         let options = ImageRequest.ThumbnailOptions(size: CGSize(width: 64, height: 64), unit: .pixels)
 
         measure {
@@ -66,9 +74,10 @@ class ImageProcessingPerformanceTests: XCTestCase {
     }
 
     // Should be roughly identical to the flexible target size.
-    func testCreateThumbnailStaticSize() throws {
-        let image = try XCTUnwrap(makeHighResolutionImage())
-        let data = try XCTUnwrap(ImageEncoders.ImageIO(type: .jpeg).encode(image))
+    @Test
+    func createThumbnailStaticSize() throws {
+        let image = try #require(makeHighResolutionImage())
+        let data = try #require(ImageEncoders.ImageIO(type: .jpeg).encode(image))
         let options = ImageRequest.ThumbnailOptions(maxPixelSize: 64)
 
         measure {