+++ Tests/NukePerformanceTests/LazyImageViewPerformanceTests.swift
@@ -2,14 +2,16 @@
 //
 // Copyright (c) 2015-2026 Alexander Grebenyuk (github.com/kean).
 
+#if !os(watchOS)
+
 import Testing
 import Foundation
 import Nuke
-import NukeExtensions
+import NukeUI
 
 @Suite(.serialized)
 @MainActor
-final class ImageViewPerformanceTests {
+final class LazyImageViewPerformanceTests {
     private let dummyCacheRequest = ImageRequest(url: URL(string: "http://test.com/9999999)")!, processors: [ImageProcessors.Resize(size: CGSize(width: 2, height: 2))])
 
     init() {
@@ -21,24 +23,22 @@ final class ImageViewPerformanceTests {
         ImagePipeline.shared.configuration.imageCache?[dummyCacheRequest] = nil
     }
 
-    // This is the primary use case that we are optimizing for - loading images
-    // into target, the API that majoriy of the apps are going to use.
     @Test
-    func imageViewMainThreadPerformance() {
-        let view = _ImageView()
+    func lazyImageViewMainThreadPerformance() {
+        let view = LazyImageView()
 
         let urls = (0..<20_000).map { _ in URL(string: "http://test.com/1)")! }
 
         measure {
             for url in urls {
-                NukeExtensions.loadImage(with: url, into: view)
+                view.url = url
             }
         }
     }
 
     @Test
-    func imageViewMainThreadPerformanceCacheHit() {
-        let view = _ImageView()
+    func lazyImageViewMainThreadPerformanceCacheHit() {
+        let view = LazyImageView()
 
         let requests = (0..<50_000).map { _ in ImageRequest(url: URL(string: "http://test.com/1)")!) }
         for request in requests {
@@ -47,36 +47,36 @@ final class ImageViewPerformanceTests {
 
         measure {
             for request in requests {
-                NukeExtensions.loadImage(with: request, into: view)
+                view.request = request
             }
         }
     }
 
     @Test
-    func imageViewMainThreadPerformanceWithProcessor() {
-        let view = _ImageView()
+    func lazyImageViewMainThreadPerformanceWithProcessor() {
+        let view = LazyImageView()
 
         let urls = (0..<20_000).map { _ in URL(string: "http://test.com/1)")! }
 
         measure {
             for url in urls {
-                let request = ImageRequest(url: url, processors: [ImageProcessors.Resize(size: CGSize(width: 1, height: 1))])
-                NukeExtensions.loadImage(with: request, into: view)
+                view.request = ImageRequest(url: url, processors: [ImageProcessors.Resize(size: CGSize(width: 1, height: 1))])
             }
         }
     }
 
     @Test
-    func imageViewMainThreadPerformanceWithProcessorAndSimilarImageInCache() {
-        let view = _ImageView()
+    func lazyImageViewMainThreadPerformanceWithProcessorAndSimilarImageInCache() {
+        let view = LazyImageView()
 
         let urls = (0..<20_000).map { _ in URL(string: "http://test.com/9999999)")! }
 
         measure {
             for url in urls {
-                let request = ImageRequest(url: url, processors: [ImageProcessors.Resize(size: CGSize(width: 1, height: 1))])
-                NukeExtensions.loadImage(with: request, into: view)
+                view.request = ImageRequest(url: url, processors: [ImageProcessors.Resize(size: CGSize(width: 1, height: 1))])
             }
         }
     }
 }
+
+#endif